[추가] 상인 경험치, 인벤토리, 영웅 정보 창 작업 중

+++ Assets/Scripts/Data/DataLv.cs
@@ -21,7 +21,7 @@ public class DataLv : IRegistrable
             MaxLv = lv;
     }
 
-    public static int GetLv(int exp)
+    public static short GetLv(int exp)
     {
         foreach(var kvp in dictLvs)
         {
@@ -32,7 +32,7 @@ public class DataLv : IRegistrable
         return 1;
     }
 
-    public static int GetTraderLv(int exp)
+    public static short GetTraderLv(int exp)
     {
         foreach (var kvp in dictLvs)
         {
